Documentation updates:
global: Modified the unicode in |stub| to be compatible with sphinx. Still appears as a wrench in the browser.

events: added information for events, and a list of events, as well as their arg paramater. Still marked as 'WIP', as I feel my documentation was a bit lacking

permissions: added information for channel\server permissions, dual channel permissions, and roles. need to update with bits at some point.