Fix #320
I'm not entirely sure this is the proper way to fix this. However, it seems like
SendFileAsync returns the Task<RestUserMessage>, which would exit the using
block, disposing of the FileStream potentially before the Task is finished.

Changed to await the result of the task before exiting the using block.