Initial audit logs implementation
Frankly, I have no idea if any of this is in the correct place, or
follows the "norm" for internal stuff.

Feedback would be nice for this implementation, as I believe this
implementation uses the least amount of partial objects (I couldn't
really avoid them in RestAuditLogEntry, unless I wanted to write entry
types for each type of entry) and is also fairly simple.