fix: strip trailing slash from ratelimit bucket IDs
This resolves #1125.

fixes a bug where some ratelimit buckets would include a trailing slash,
while others wouldn't; this would cause them to be treated as separate
ratelimits, even though they are the same

ideally this fix should change the ratelimit generator, but that code is
pretty complicated and this was an easier fix that seems less likely to
break things in the future.

tested against normal bot function, all routes are assigned the proper
buckets from my testing, so this should be good to go.