Add NestedChannelProperties type
This change fixes an issue that was introduced with the INestedChannel type. Previously, it was possible to use ModifyAsync on a ChannelCategory to set it's CategoryId. This is not supported, and would probably error.
This is a breaking change, however I can't imagine that anyone was trying to nest categories before this. I don't think that it should affect anyone.

This change adds the NestedChannelProperties and ModifyNestedChannelParams types, which follow the same class hierarchy as the public-facing channel entities. This allows only guild voice and text channels to be placed under a category.
The change