Allow for custom default RequestOptions
This change would allow the consumer of the library to set the default `RequestOptions` if they desire.  Invoking a method that takes `RequestOptions` as a parameter without specifying the argument will use a copy of the default.  Behavior will be unchanged without modifying the `RequestOptions.Default` property.

Additionally, this change adds 4 `With...` methods that allow for fluent construction of new `RequestOptions` objects.