Added ClientId and ClientSecret to the library. These can be left null without any effect to the rest of the library. Just don't expect the refresh methods to work.
Added SendFormAsync in order to send url encoded content. It uses the bucket system all the oder SendAsync methods use.

Added a GetTokenAsync method. This method  is used to get either a token from a code, or a token from a refresh token. It raises an error if given any other kind of tokens.