Cache unknown thread channel on THREAD_UPDATE (#136)
In the case where a thread is archived upon login, the thread channel is
not cached by the client. When the thread channel is unarchived, the
library simply outputs an unknown channel warning. When messages are
sent in the unarchived thread, the library won't have it cached and
will only output warnings.

This change makes it so the channel is cached when unarchived and
ThreadUpdated is invoked with an non specified "before" parameter. From
that point, the channel will be available for future events.